remove min, flatten structure, build via npm scripts
